requirements.txt
requests==2.28.1
Mastodon.py==1.5.2
